requirements.txt
requests==2.31.0
pandas>=1.5.0
pymongo==4.6.0
python-dotenv==1.0.0
eurostat==1.0.4
numpy==1.24.3
tqdm==4.66.1
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
motor>=3.0.0
plotly>=5.10.0
fastapi>=0.68.0
uvicorn>=0.15.0
jinja2>=3.0.0
python-multipart>=0.0.5
